[게시판 회원 기능] 티버 미션 제출합니다. (#76)
* feat: add gradle dependency

* feat: get/post practice

* refactor: modify path

* fix: add responseBody annotation

* feat: click the button to move article-edit page

* test: modify uri

* feat: create and save article

* feat: write the article

* feat: look up the article

* feat: look up the article by id

* fix: find an article by id

* feat: update the article

* feat: delete the article

* feat: move to index from article-edit

* feat: setting devtools

* fix : UI bug

* refactor: remove duplicated html code

* refactor: remove duplicated index logo

* refactor: create ArticleService

* refactor: divide package

* refactor: divide package

* refactor: modify return value

* refactor: modify test cod

- fix bug
- extract duplicate fields

* feat: update README.md

* refactor: update README.md

* refactor: remove MyblogApplication class

* refactor: modify newline

* refactor: modify return value of save method

* refactor: add ArticleService field

- remove ArticleRepository field

* refactor: for convention

* refactor: create NotFoundArticleException

* refactor: modify field name

* refactor: modify newline

* refactor: extract duplicate code

* refactor: create matchId method

* refactor: modify return value of findAll method

* refactor: modify annotation

* refactor: create update method in Article class

* refactor: modify method name, location

* test: create ArticleRepositoryTests

* refactor: modify field type of id

* refactor: remove setter of title, contents, coverUrl field

* refactor: modify semicolon position

* refactor: modify URL

* test: modify findAll method

* test: modify uri

* test: add ability to check location of header

* refactor: remove ArticleService class

* refactor: modify path

* refactor: modify id field type

* refactor: modify injection

* refactor: fix bug

* refactor: for code convention

* refactor: modify properties

* refactor: add spring-data-jap and h2

- create ArticleDto

* feat: mysql 연동

* docs: 기능 목록 추가

* feat: MySQL 사용 설정

* fix: 템플릿 파일명 중복제거 및 오타 수정

* feat: signup get 구현 및 테스트 추가

* add: User 및 UserDto 클래스 추가

* feat: users post 데이터 받아오는 기능 구현

* feat: 이름 및 패스워드 규칙 추가

이름은 최소 2글자, 최대 10글자까지 입력 가능.
패스워드는 최소 8글자 이상으로 알파벳 소문자, 대문자, 숫자, 특수문자의 조합이다.

* feat: 프론트엔드에 입력 검증 기능 추가

* feat: 비밀번호 확인 기능 동작 추가

* feat: 회원가입 실패 시 에러 메시지 페이지에 노출

* feat: 동일한 email 회원가입 불가 기능 구현

* test: 회원등록 기능 테스트 구현

* feat: 회원조회 기능 구현

* test: 회원조회 기능 테스트 구현

* feat: 로그인 기능 구현

- 로그인 성공 시 메인 화면을 띄우고 우측 상단에 사용자 이름을 띄운다.
- 로그인 실패 시 상황에 맞는 실패 메시지를 띄운다.

* fix: 세션 생성 방법 수정

* test: 로그인 기능 테스트 구현

* feat: 로그아웃 기능 구현

* test: 로그아웃 기능 테스트 구현

* refactor: 로그인 시 세션 관리 수정

* test: 로그인 기능 테스트 수정

* feat: 로그인 한 유저가 로그인/회원가입 화면에 접근할 경우 메인 화면으로 이동

* refactor: 로그아웃 시 세션 처리 방법 수정

* refactor: 컨트롤러명 수정

* test: UserControllerTest 테스트 코드 수정

* feat: mypage, mypage-edit 접근 권한 기능 추가

- session확인 시 email로 확인하도록 수정

* test: 로그인하지 않았을 경우 mypage, mypage-edit 이동 실패 테스트 구현

- 쿠키에서 session 확인하도록 LoginControllerTests 메소드 수정

* feat: 회원 탈퇴 기능 구현

* refactor: 회원 탈퇴 시 DeleteMapping으로 수정

* feat: 회원 수정 기능 구현

* docs: 기능 목록 수정

* refactor: dto.request 패키지 생성

* test: test 코드 중복 제거

- LogoutControllerTests 삭제
- session확인 코드 삭제

* refactor: 코드 컨벤션을 위한 수정

* refactor: properties 설정 파일 수정

* feat: siguup 페이지에서 login 페이지로 이동하도록 url 수정

* fix: git conflict 해결 과정 중 발생한 오류 수정

* refactor: ArticleDto.valueOfArticle 메소드 구현

* test: ArticleDto.valueOfArticle 메소드 테스트 구현

* refactor: 존재하지 않는 Article 요청 시 error 메세지 반환

* refactor: valueOfArticle 메소드 오버로딩

* refactor: 변수명 변경

* fix: ArticleDto 관련 버그 수정

* refactor: confirmSession 메소드 생성

* test: User의 패스워드가 일치하지 않는 경우 테스트 코드 구현

* test: 회원가입 실패 시 테스트 케이스 추가

* fix: LoginController의 메소드 반환 값 수정

* test: 테스트 코드 중복 제거

* refactor: 변수명 수정

* refactor: User 생성자 수정

* refactor: LoginService 클래스 분리

- Login, NotFoundUserEmail, NotFoundMatchPassword 예외 클래스 구현
- GlobalExceptionHandler 클래스 구현

* test: Login 실패 시 테스트 코드 수정

* refactor: 존재하지 않는 게시글에 대한 예외 처리 수정

* test: 존재하지 않는 게시글에 대한 테스트 코드 구현

* refactor: service클래스 분리

* feat: BasicAuthInterceptor, WebMvcConfig 클래스 구현

* test: 로그인 되어있는 경우 로그인 실패하는 테스트 구현

* feat: Interceptor 구현

* test: Interceptor 테스트 코드 구현

* refactor: 메소드명 변경, 반환 값 수정

* refactor: 메소드명 변경, 반환 값 수정

* refactor: 코드 컨벤션을 위한 리팩토링

* refactor: Interceptor 와 config 수정

* refactor: confirmSession 수정

* refactor: session 관련 테스트 코드 구현

* refactor: 코드 컨벤션을 위한 리팩토링

* refactor: GlobalExceptionHandler 클래스 이동

+++ src/main/java/techcourse/myblog/web/ArticleController.java
@@ -1,6 +1,10 @@
 package techcourse.myblog.web;
 
+import javax.servlet.http.HttpSession;
+
 import techcourse.myblog.domain.Article;
+import techcourse.myblog.dto.request.ArticleDto;
+import techcourse.myblog.exception.NotFoundArticleException;
 import techcourse.myblog.repository.ArticleRepository;
 
 import org.springframework.stereotype.Controller;
@@ -21,42 +25,66 @@ public class ArticleController {
 		return "index";
 	}
 
-	@GetMapping("writing")
-	public String createArticle() {
+	@GetMapping("/writing")
+	public String createArticle(HttpSession httpSession) {
+		if (!existSession(httpSession)) {
+			return "redirect:/";
+		}
 		return "article-edit";
 	}
 
-	@PostMapping("articles")
-	public String saveArticle(Article article, Model model) {
-		model.addAttribute(article);
-		articleRepository.save(article);
-		return "article";
+	@PostMapping("/articles")
+	public String saveArticle(ArticleDto articleDto, HttpSession httpSession) {
+		if (!existSession(httpSession)) {
+			return "redirect:/";
+		}
+		Article article = articleDto.valueOfArticle();
+		Long id = articleRepository.save(article).getId();
+		return "redirect:/articles/" + id;
 	}
 
-	@GetMapping("articles/{articleId}")
+	@GetMapping("/articles/{articleId}")
 	public String getArticle(@PathVariable Long articleId, Model model) {
-		model.addAttribute("article", articleRepository.findById(articleId));
+		getArticleWhenExists(articleId, model);
 		return "article";
 	}
 
-	@GetMapping("articles/{articleId}/edit")
-	public String editArticle(@PathVariable Long articleId, Model model) {
-		model.addAttribute("article", articleRepository.findById(articleId));
+	@GetMapping("/articles/{articleId}/edit")
+	public String editArticle(@PathVariable Long articleId, Model model, HttpSession httpSession) {
+		if (!existSession(httpSession)) {
+			return "redirect:/";
+		}
+		getArticleWhenExists(articleId, model);
 		return "article-edit";
 	}
 
-	@PutMapping("articles/{articleId}")
-	public String modifyArticle(@PathVariable Long articleId, Article article, Model model) {
-		article.setId(articleId);
-		articleRepository.update(article);
-		model.addAttribute(articleRepository.findById(articleId));
-		return "article";
+	private void getArticleWhenExists(Long articleId, Model model) {
+		Article article = articleRepository.findById(articleId)
+				.orElseThrow(NotFoundArticleException::new);
+		model.addAttribute("article", article);
 	}
 
-	@DeleteMapping("articles/{articleId}")
-	public String deleteArticle(@PathVariable Long articleId) {
+	@PutMapping("/articles/{articleId}")
+	public String modifyArticle(@PathVariable Long articleId, ArticleDto articleDto, HttpSession httpSession) {
+		if (!existSession(httpSession)) {
+			return "redirect:/";
+		}
+		Article article = articleDto.valueOfArticle(articleId);
+		articleRepository.save(article);
+		return "redirect:/articles/" + articleId;
+	}
+
+	@DeleteMapping("/articles/{articleId}")
+	public String deleteArticle(@PathVariable Long articleId, HttpSession httpSession) {
+		if (!existSession(httpSession)) {
+			return "redirect:/";
+		}
 		articleRepository.deleteById(articleId);
 		return "redirect:/";
+	}
+
+	private boolean existSession(HttpSession httpSession) {
+		return (httpSession.getAttribute("email") != null);
 	}
 }
